requirements.txt
streamlit
boto3
folium
python-dotenv
